[WasmFS] Allow backends to have short reads and writes
We previously tried to prevent reads and writes beyond the ends of files by
calling `getSize` on files in the backend-independent implementations of `read`
and `write` and adjusting the length we passed to the backend accordingly. This
works fine when files have meaningful lengths, reads and writes are guaranteed
not to be shortened by the underlying APIs, and files cannot change size
concurrently with the syscall execution, but those requirements are not always
met.

Instead, change the backend API to match the `pread` and `pwrite` syscall APIs,
which allow accesses to be shorter than requested. Specifically, change the
return types to be `ssize_t` so that the accessed lengths can be returned. Error
codes are now returned as negative values, just like in the syscalls themselves.

This change will also have performance benefits, especially for the OPFS
backend, in which getting the size of a file is an asynchronous operation that
is much slower than a small read or write.

+++ src/library_wasmfs_jsimpl.js
@@ -84,7 +84,7 @@ mergeInto(LibraryManager.library, {
     assert(wasmFS$backends[backend]);
 #endif
     var result = await wasmFS$backends[backend].write(file, buffer, length, offset);
-    {{{ makeSetValue('result_p', 0, 'result', 'i64') }}};
+    {{{ makeSetValue('result_p', 0, 'result', 'i32') }}}; // TODO: wasm64
     _emscripten_proxy_finish(ctx);
   },
 
@@ -95,7 +95,7 @@ mergeInto(LibraryManager.library, {
     assert(wasmFS$backends[backend]);
 #endif
     var result = await wasmFS$backends[backend].read(file, buffer, length, offset);
-    {{{ makeSetValue('result_p', 0, 'result', 'i32') }}};
+    {{{ makeSetValue('result_p', 0, 'result', 'i32') }}}; // TODO: wasm64
     _emscripten_proxy_finish(ctx);
   },
